split nodes, expansions, utils files
