+++ src/gtc/dace/oir_to_dace.py
@@ -23,109 +23,7 @@ from gtc.oir import (
     VerticalLoopSection,
 )
 from gtc.passes.oir_optimizations.utils import AccessCollector
-
-
-class CartesianIJIndexSpace(tuple):
-    def __new__(cls, *args, **kwargs):
-        return super(CartesianIJIndexSpace, cls).__new__(cls, *args, **kwargs)
-
-    def __init__(self, *args, **kwargs):
-        msg = "CartesianIJIndexSpace must be a pair of pairs of integers."
-        if not len(self) == 2:
-            raise ValueError(msg)
-        if not all(len(v) == 2 for v in self):
-            raise ValueError(msg)
-        if not all(isinstance(v[0], int) and isinstance(v[1], int) for v in self):
-            raise ValueError(msg)
-
-    @classmethod
-    def from_offset(
-        cls, offset: Union[Tuple[int, ...], CartesianOffset]
-    ) -> "CartesianIJIndexSpace":
-        if isinstance(offset, CartesianOffset):
-            return cls((offset.i, offset.i), (offset.j, offset.j))
-        return cls(((offset[0], offset[0]), (offset[1], offset[1])))
-
-    @classmethod
-    def from_iteration_space(
-        cls, iteration_space: CartesianIterationSpace
-    ) -> "CartesianIJIndexSpace":
-        return cls(
-            (
-                (iteration_space.i_interval.start.offset, iteration_space.i_interval.end.offset),
-                (iteration_space.j_interval.start.offset, iteration_space.j_interval.end.offset),
-            )
-        )
-
-    def compose(
-        self, other: Union["CartesianIterationSpace", "CartesianIJIndexSpace"]
-    ) -> "CartesianIJIndexSpace":
-        if isinstance(other, CartesianIterationSpace):
-            other = CartesianIJIndexSpace.from_iteration_space(other)
-        return CartesianIJIndexSpace(
-            (
-                (self[0][0] + other[0][0], (self[0][1] + other[0][1])),
-                (self[1][0] + other[1][0], (self[1][1] + other[1][1])),
-            )
-        )
-
-    def __or__(
-        self, other: Union["CartesianIterationSpace", "CartesianIJIndexSpace"]
-    ) -> "CartesianIJIndexSpace":
-        if isinstance(other, CartesianIterationSpace):
-            other = CartesianIJIndexSpace.from_iteration_space(other)
-        return CartesianIJIndexSpace(
-            (
-                (min(self[0][0], other[0][0]), max(self[0][1], other[0][1])),
-                (min(self[1][0], other[1][0]), max(self[1][1], other[1][1])),
-            )
-        )
-
-
-def nodes_extent_calculation(
-    nodes: Collection[Union[VerticalLoopLibraryNode, HorizontalExecutionLibraryNode]]
-) -> Dict[str, Tuple[Tuple[int, int], Tuple[int, int]]]:
-    access_spaces: Dict[str, Tuple[Tuple[int, int], ...]] = dict()
-    inner_nodes = []
-    for node in nodes:
-        if isinstance(node, VerticalLoopLibraryNode):
-            for _, section_sdfg in node.sections:
-                for he in (
-                    ln
-                    for ln, _ in section_sdfg.all_nodes_recursive()
-                    if isinstance(ln, HorizontalExecutionLibraryNode)
-                ):
-                    inner_nodes.append(he)
-        else:
-            assert isinstance(node, HorizontalExecutionLibraryNode)
-            inner_nodes.append(node)
-    for node in inner_nodes:
-        access_collection = AccessCollector.apply(node.oir_node)
-        iteration_space = node.oir_node.iteration_space
-        if iteration_space is not None:
-            for name, offsets in access_collection.offsets().items():
-                for off in offsets:
-                    access_extent = (
-                        (
-                            iteration_space.i_interval.start.offset + off[0],
-                            iteration_space.i_interval.end.offset + off[0],
-                        ),
-                        (
-                            iteration_space.j_interval.start.offset + off[1],
-                            iteration_space.j_interval.end.offset + off[1],
-                        ),
-                    )
-                    if name not in access_spaces:
-                        access_spaces[name] = access_extent
-                    access_spaces[name] = tuple(
-                        (min(asp[0], ext[0]), max(asp[1], ext[1]))
-                        for asp, ext in zip(access_spaces[name], access_extent)
-                    )
-
-    return {
-        name: ((-asp[0][0], asp[0][1]), (-asp[1][0], asp[1][1]))
-        for name, asp in access_spaces.items()
-    }
+from gtc.dace.utils import nodes_extent_calculation, CartesianIJIndexSpace
 
 
 class BaseOirSDFGBuilder(ABC):
