+++ tests/test_unittest/test_gtc/test_dace.py
@@ -11,9 +11,9 @@ from gt4py.backend.gtc_backend.defir_to_gtir import DefIRToGTIR
 from gt4py.definitions import BuildOptions
 from gt4py.frontend.gtscript_frontend import GTScriptFrontend
 from gtc.dace.nodes import HorizontalExecutionLibraryNode, VerticalLoopLibraryNode
-from gtc.dace_to_oir import convert
+from gtc.dace.dace_to_oir import convert
 from gtc.gtir_to_oir import GTIRToOIR, oir_iteration_space_computation
-from gtc.oir_to_dace import OirSDFGBuilder
+from gtc.dace.oir_to_dace import OirSDFGBuilder
 from gtc.passes.gtir_pipeline import GtirPipeline
 from gtc.passes.oir_optimizations.caches import (
     IJCacheDetection,
